refactor(redirect-endpoint): reorganize tests for MVC pattern
