+++ src/test/java/tech/brokendeveloper/shortener_url/dto/v2/ShortenUrlRequestDtoV2Test.java
@@ -1,10 +1,9 @@
-package tech.brokendeveloper.shortener_url.domain.url.api.v2.dto;
+package tech.brokendeveloper.shortener_url.dto.v2;
 
 import jakarta.validation.Validation;
 import jakarta.validation.Validator;
 import jakarta.validation.ValidatorFactory;
 import org.junit.jupiter.api.Test;
-import tech.brokendeveloper.shortener_url.dto.v2.ShortenUrlRequestDtoV2;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
