+++ src/test/java/tech/brokendeveloper/shortener_url/services/GenerateShortUrlServiceV2Test.java
@@ -1,4 +1,4 @@
-package tech.brokendeveloper.shortener_url.domain.url.useCases;
+package tech.brokendeveloper.shortener_url.services;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -11,7 +11,6 @@ import tech.brokendeveloper.shortener_url.dto.v2.ShortenUrlResponseDtoV2;
 import tech.brokendeveloper.shortener_url.domain.Url;
 import tech.brokendeveloper.shortener_url.repositories.UrlRepository;
 import tech.brokendeveloper.shortener_url.exceptions.ShortUrlGenerationException;
-import tech.brokendeveloper.shortener_url.services.GenerateShortUrlServiceV2;
 import tech.brokendeveloper.shortener_url.utils.UrlBuilder;
 import tech.brokendeveloper.shortener_url.utils.NanoIdShortCodeGenerator;
